refactor: move symlink related functions from root.c -> root/symlink.c
