+++ Makefile.am
@@ -71,6 +71,7 @@ watchman_SOURCES = \
 	main.c       \
 	root.c       \
 	root/lock.c       \
+	root/symlink.c       \
 	state.c      \
 	string.c     \
 	time.c
